마크업 수정 + 컴포넌트 분리
